feat: 해당역 디자인 조회(인기순) 마지막 데이터 조건 추가 - #38

+++ cakey-api/src/main/java/com/cakey/storelikes/dto/StorePopularityLikedByUserRes.java
@@ -8,13 +8,13 @@ import java.util.List;
 
 @Builder(access = AccessLevel.PRIVATE)
 public record StorePopularityLikedByUserRes(
-        long nextLikesCursor,
+        Integer nextLikesCursor,
         Long nexStoreId,
         long storeCount,
         List<StoreInfo> stores
 ) {
-    public static StorePopularityLikedByUserRes of(final long nextLikesCursor,
-                                                   final long nexStoreId,
+    public static StorePopularityLikedByUserRes of(final Integer nextLikesCursor,
+                                                   final Long nexStoreId,
                                                    final long storeCount,
                                                    final List<StoreInfo> stores) {
         return StorePopularityLikedByUserRes.builder()
